+++ setup.py
@@ -20,7 +20,7 @@ setup(
                    'hpy/devel/src/runtime/ctx_module.c',
                    'hpy/devel/src/runtime/ctx_type.c',
                    'hpy/devel/src/runtime/argparse.c',
-                   'hpy/devel/src/runtime/listbuilder.c',
+                   'hpy/devel/src/runtime/ctx_listbuilder.c',
                   ],
                   include_dirs=[
                       'hpy/devel/include',
